requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
fastapi==0.118.0
greenlet==3.2.4
idna==3.10
passlib==1.7.4
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
